Revert "more coverage on test"
This reverts commit b2112d16afb9b88807314932c427f31a968e41c3.
