+++ app/src/main/java/com/swent/assos/config/Config.kt
@@ -1,8 +1,5 @@
 package com.swent.assos.config
 
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.functions.FirebaseFunctions
 import java.net.HttpURLConnection
 import java.net.URL
 import org.json.JSONObject
@@ -77,27 +74,5 @@ class Config {
 
   fun get_demo(callback: (Boolean) -> Unit) {
     checkFirebaseEmulator("10.0.2.2", "9099") { isConnected -> callback(isConnected) }
-  }
-
-  fun init() {
-    this.get_all { onlineServices ->
-      val firestoreEmu = onlineServices.contains("firestore")
-      val authEmu = onlineServices.contains("auth")
-      if (firestoreEmu) {
-        // Configure Firestore to use the Firestore emulator
-        if (FirebaseFirestore.getInstance().firestoreSettings.host != "10.0.2.2:8080") {
-          FirebaseFirestore.getInstance().useEmulator("10.0.2.2", 8080)
-          if (authEmu) {
-            // Configure Firebase Auth to use the Auth emulator
-            FirebaseAuth.getInstance().useEmulator("10.0.2.2", 9099)
-          }
-
-          val emu = onlineServices.contains("functions")
-          if (emu) {
-            FirebaseFunctions.getInstance().useEmulator("10.0.2.2", 5001)
-          }
-        }
-      }
-    }
   }
 }
